Crate mp4parse

Module for parsing ISO Base Media Format aka video/mp4 streams.

Modules

boxes

Structs

AudioSampleEntry

BMFFBox

BoxHeader
Basic ISO box structure.

BoxIter

ChannelMappingTable

ChunkOffsetBox

Edit

EditListBox
Edit list box 'elst'

EmptySampleTableBoxes
A fragmented file contains no sample data in stts, stsc, and stco.

FileTypeBox
File type box 'ftyp'.

HandlerBox

MediaContext
Internal data structures.

MediaHeaderBox
Media header box 'mdhd'

MediaScaledTime
A time to be scaled by the media's global (mvhd) timescale.

MediaTimeScale
The media's global (mvhd) timescale in units per second.

MovieHeaderBox
Movie header box 'mvhd'.

OpusSpecificBox
Represent an OpusSpecificBox 'dOps'

Sample

SampleDescriptionBox

SampleSizeBox

SampleToChunk

SampleToChunkBox

SyncSampleBox

TimeToSampleBox

Track

TrackHeaderBox
Track header box 'tkhd'

TrackScaledTime
A time to be scaled by the track's local (mdhd) timescale. Members are time in scale units and the track id.

TrackTimeScale
The track's local (mdhd) timescale. Members are timescale units per second and the track id.

VPxConfigBox
Represent a Video Partition Codec Configuration 'vpcC' box (aka vp9).

VideoSampleEntry

Enums

AudioCodecSpecific

Error
Describes parser failures.

SampleEntry

TrackType

VideoCodecSpecific

Constants

BUF_SIZE_LIMIT

Statics

DEBUG_MODE

Functions

be_i16

be_i32

be_i64

be_u16

be_u32

be_u64

get_debug_mode

parse_mdhd

parse_mvhd

read_audio_desc
Parse an audio description inside an stsd box.

read_box_header
Read and parse a box header.

read_buf
Read size bytes into a Vector or return error.

read_co64
Parse a co64 box.

read_dops
Parse OpusSpecificBox.

read_edts

read_elst
Parse a elst box.

read_fixed_length_pascal_string

read_ftyp
Parse an ftyp box.

read_fullbox_extra
Parse the extra header fields for a full box.

read_hdlr
Parse a hdlr box.

read_mdhd
Parse a mdhd box.

read_mdia

read_minf

read_moov

read_mp4
Read the contents of a box, including sub boxes.

read_mvhd
Parse an mvhd box.

read_null_terminated_string

read_pascal_string

read_stbl

read_stco
Parse a stco box.

read_stsc
Parse a stsc box.

read_stsd
Parse a stsd box.

read_stss
Parse a stss box.

read_stsz
Parse a stsz box.

read_stts
Parse a stts box.

read_tkhd
Parse a tkhd box.

read_trak

read_video_desc
Parse an video description inside an stsd box.

read_vpcc
Parse a VPx Config Box.

serialize_opus_header
Re-serialize the Opus codec-specific config data as an OpusHead packet.

set_debug_mode

skip
Skip a number of bytes that we don't care to parse.

skip_box_content
Skip over the entire contents of a box.

Type Definitions

Result
Result shorthand using our Error enum.