Enum mp4parse::Error

pub enum Error {
    InvalidData(&'static str),
    Unsupported(&'static str),
    UnexpectedEOF,
    Io(Error),
    NoMoov,
}

Describes parser failures.

This enum wraps the standard io::Error type, unified with our own parser error states and those of crates we use.

Variants
InvalidData(&'static str)

Parse error caused by corrupt or malformed data.

Unsupported(&'static str)

Parse error caused by limited parser support rather than invalid data.

UnexpectedEOF

Reflect std::io::ErrorKind::UnexpectedEof for short data.

Io(Error)

Propagate underlying errors from std::io.

NoMoov

read_mp4 terminated without detecting a moov box.