Crate mp4parse

Module for parsing ISO Base Media Format aka video/mp4 streams.

Structs

AudioSampleEntry

ES_Descriptor

EmptySampleTableBoxes
A fragmented file contains no sample data in stts, stsc, and stco.

FLACMetadataBlock

FLACSpecificBox
Represet a FLACSpecificBox 'dfLa'

MediaContext
Internal data structures.

MediaScaledTime
A time to be scaled by the media's global (mvhd) timescale.

MediaTimeScale
The media's global (mvhd) timescale in units per second.

MovieExtendsBox

OpusSpecificBox
Represent an OpusSpecificBox 'dOps'

ProtectionSchemeInfoBox

ProtectionSystemSpecificHeaderBox

Track

TrackEncryptionBox

TrackHeaderBox
Track header box 'tkhd'

TrackScaledTime
A time to be scaled by the track's local (mdhd) timescale. Members are time in scale units and the track id.

TrackTimeScale
The track's local (mdhd) timescale. Members are timescale units per second and the track id.

VPxConfigBox
Represent a Video Partition Codec Configuration 'vpcC' box (aka vp9).

VideoSampleEntry

Enums

AudioCodecSpecific

CodecType

Error
Describes parser failures.

SampleEntry

TrackType

VideoCodecSpecific

Functions

read_mp4
Read the contents of a box, including sub boxes.

serialize_opus_header
Re-serialize the Opus codec-specific config data as an OpusHead packet.

set_debug_mode

Type Definitions

ByteData

Result
Result shorthand using our Error enum.